refactor[venom]: refactor module structure (#4295)
this commit refactors venom's analysis/ and passes/ module structure so
that you can import analyses and passes directly from
`vyper.venom.analysis` and `vyper.venom.passes` directly instead of
needing to know the module names.

this commit also removes the `stack_reorder.py` file, which is dead code.
