+++ tests/unit/compiler/venom/test_stack_reorder.py
@@ -1,7 +1,7 @@
 from vyper.venom import generate_assembly_experimental
-from vyper.venom.analysis.analysis import IRAnalysesCache
+from vyper.venom.analysis import IRAnalysesCache
 from vyper.venom.context import IRContext
-from vyper.venom.passes.store_expansion import StoreExpansionPass
+from vyper.venom.passes import StoreExpansionPass
 
 
 def test_stack_reorder():
