+++ vyper/venom/__init__.py
@@ -9,16 +9,18 @@ from vyper.venom.analysis.analysis import IRAnalysesCache
 from vyper.venom.context import IRContext
 from vyper.venom.function import IRFunction
 from vyper.venom.ir_node_to_venom import ir_node_to_venom
-from vyper.venom.passes.algebraic_optimization import AlgebraicOptimizationPass
-from vyper.venom.passes.branch_optimization import BranchOptimizationPass
-from vyper.venom.passes.dft import DFTPass
-from vyper.venom.passes.make_ssa import MakeSSA
-from vyper.venom.passes.mem2var import Mem2Var
-from vyper.venom.passes.remove_unused_variables import RemoveUnusedVariablesPass
-from vyper.venom.passes.sccp import SCCP
-from vyper.venom.passes.simplify_cfg import SimplifyCFGPass
-from vyper.venom.passes.store_elimination import StoreElimination
-from vyper.venom.passes.store_expansion import StoreExpansionPass
+from vyper.venom.passes import (
+    SCCP,
+    AlgebraicOptimizationPass,
+    BranchOptimizationPass,
+    DFTPass,
+    MakeSSA,
+    Mem2Var,
+    RemoveUnusedVariablesPass,
+    SimplifyCFGPass,
+    StoreElimination,
+    StoreExpansionPass,
+)
 from vyper.venom.venom_to_assembly import VenomCompiler
 
 DEFAULT_OPT_LEVEL = OptimizationLevel.default()
