+++ vyper/venom/analysis/__init__.py
@@ -1,0 +1,6 @@
+from .analysis import IRAnalysesCache, IRAnalysis
+from .cfg import CFGAnalysis
+from .dfg import DFGAnalysis
+from .dominators import DominatorTreeAnalysis
+from .equivalent_vars import VarEquivalenceAnalysis
+from .liveness import LivenessAnalysis
