+++ vyper/venom/passes/__init__.py
@@ -1,0 +1,11 @@
+from .algebraic_optimization import AlgebraicOptimizationPass
+from .branch_optimization import BranchOptimizationPass
+from .dft import DFTPass
+from .make_ssa import MakeSSA
+from .mem2var import Mem2Var
+from .normalization import NormalizationPass
+from .remove_unused_variables import RemoveUnusedVariablesPass
+from .sccp import SCCP
+from .simplify_cfg import SimplifyCFGPass
+from .store_elimination import StoreElimination
+from .store_expansion import StoreExpansionPass
