+++ vyper/venom/passes/algebraic_optimization.py
@@ -1,5 +1,4 @@
-from vyper.venom.analysis.dfg import DFGAnalysis
-from vyper.venom.analysis.liveness import LivenessAnalysis
+from vyper.venom.analysis import DFGAnalysis, LivenessAnalysis
 from vyper.venom.basicblock import IRInstruction, IRLabel, IRLiteral, IROperand
 from vyper.venom.passes.base_pass import IRPass
