+++ vyper/venom/passes/remove_unused_variables.py
@@ -1,6 +1,5 @@
 from vyper.utils import OrderedSet
-from vyper.venom.analysis.dfg import DFGAnalysis
-from vyper.venom.analysis.liveness import LivenessAnalysis
+from vyper.venom.analysis import DFGAnalysis, LivenessAnalysis
 from vyper.venom.basicblock import IRInstruction
 from vyper.venom.passes.base_pass import IRPass
